Fix error in computation of SLS quasi-permanent combinations
The permanent loads alone (without any variable load) were missing
among those combinations. The error was sometimes masked by another
error in the code: for those quasi-permanent combinations the seismic
situation flag was incorrectly set to true. Due to this error, when any
of the variable loads had a zero value for its psi_2 combination factor
the result was apparently correct.

+++ verif/tests/actions/load_combinations/ec0/test_ec0_railway_bridge_load_combinations_01.py
@@ -59,5 +59,5 @@ if comparisonOK:
    print('test '+fname+': ok.')
 else:
     lmsg.error(fname+' ERROR.')
-
-os.remove(outputPath) # clean up after yourself.
+if comparisonOK:
+    os.remove(outputPath) # clean up after yourself.
